forest_backend/requirements.txt
flask==0.12.4
flask_restful==0.3.7
flask_script==2.0.6
flask_migrate==2.4.0
marshmallow==2.19.2
flask_sqlalchemy==2.3.2
flask_marshmallow==0.10.0
marshmallow-sqlalchemy==0.16.1
psycopg2-binary==2.8
pylint-flask==0.6
APScheduler==3.6.0
requests==2.21.0